[Contractor] Add docuseal contracts into the invite flow (#14281)
## Summary of the problem
We need to have the organizer sign the form before they are able to
invite them


## Describe your changes
Integrates docuseal contracts into the flow and generates contract when
loading second page of form (voids and regenerates contract if user goes
back into details and changes them)


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

+++ app/models/contract.rb
@@ -95,11 +95,11 @@ class Contract < ApplicationRecord
           reissue_messages.each do |role, message|
             party(role)&.notify_reissued(message:) if message.present?
           end
+          notifiable_parties.each(&:schedule_reminders)
         elsif contractable.contract_notify_when_sent
           notifiable_parties.each(&:notify)
+          notifiable_parties.each(&:schedule_reminders)
         end
-
-        notifiable_parties.each(&:schedule_reminders)
       end
     end
 
@@ -289,6 +289,8 @@ class Contract < ApplicationRecord
   end
 
   def archive_on_docuseal!
+    return if external_id.blank?
+
     docuseal_client.delete("/submissions/#{external_id}")
   end
